+++ ic-management-canister-types/CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.4.1] - 2025-09-04
+
+### Fixed
+
+- Used `candid:Reserved` inside `TakenFromCanister` and `MetadataUpload` variants in `SnapshotSource`.
+- Renamed `MainMemory` to `WasmMemory` in `SnapshotDataKind` and `SnapshotDataOffset`.
+- Added `source` field to `LoadSnapshotRecord`.
+
+While this is technically a breaking change in the Rust type system, we are treating it as a patch fix.
+This is because the affected types and methods are for new, unreleased features (snapshot download/upload).
+Therefore, no existing services or canisters should be impacted by this change.
+
+## [0.4.0] - 2025-08-25
+
+### Changed
+
+- Added `environment_variable` field to `CanisterSettings` and `DefiniteCanisterSettings`.
+  - Added the type `EnvironmentVariable`.
+- Added `settings_change` variant to `ChangeDetails`.
+- Added `environment_variables_hash` field to `CreationRecord`.
+- Added `is_replicated` field to `HttpRequestArgs`.
+
 ## [0.3.3] - 2025-08-20
 
 ### Fixed